:bug: fix settings text color
